--- fcmpy/ML/output.csv
@@ -1,6 +1,0 @@
-name,b1,L,slope,h,train_error,test_error,training_time,weights
-irisnorm,1.00,0,1.00,1.00,0.010958062673625282,0.011505890810329785,0.0015982,"(array([[ 1.        , -0.08231225,  2.13497966],
-       [-0.08915972,  1.        , -1.035009  ],
-       [ 0.35466021, -0.15872987,  1.        ]]), array([[-2875.48348736],
-       [ 1470.18265665],
-       [ 1305.52325729]]))"
