+++ fcmpy/__init__.py
