+++ fcmpy/expert_fcm/expert_based_fcm.py
@@ -1,7 +1,3 @@
-import sys, os
-myPath = os.path.dirname(os.path.abspath(__file__))
-sys.path.insert(0, myPath + '/../')
-
 from fcmpy.expert_fcm.input_validator import type_check
 from fcmpy.expert_fcm.methodsStore import EntropyStore, ReaderStore, MembershipStore, ImplicationStore, AggregationStore, DefuzzStore
 from fcmpy.expert_fcm.transform import Transform
