+++ fcmpy/simulator/transfer.py
@@ -1,7 +1,3 @@
-import sys, os
-myPath = os.path.dirname(os.path.abspath(__file__))
-sys.path.insert(0, myPath + '/../')
-
 import numpy as np
 from abc import ABC, abstractmethod
