+++ setup.cfg
@@ -1,0 +1,24 @@
+[flake8]
+max-line-length = 100
+[mypy]
+warn_unused_ignores = True
+warn_redundant_casts = True
+ignore_missing_imports = True
+
+[mypy-pdoc.test.example_pkg.*]
+ignore_errors = True
+
+[coverage:run]
+source =
+    pdoc
+omit =
+    pdoc/test/example_pkg/*
+
+[coverage:report]
+exclude_lines =
+    ^\s*continue\b
+    ^\s*return\b
+    ^\s*raise\b
+    ^\s*except\b
+    ^\s*warnings\.warn\(
+    ^\s*warn\(
